+++ configuration_js.go
@@ -1,4 +1,4 @@
-// +build !js
+// +build js,wasm
 
 package webrtc
 
@@ -31,18 +31,8 @@ type Configuration struct {
 	// unless it can be successfully authenticated with the provided name.
 	PeerIdentity string
 
-	// Certificates describes a set of certificates that the PeerConnection
-	// uses to authenticate. Valid values for this parameter are created
-	// through calls to the GenerateCertificate function. Although any given
-	// DTLS connection will use only one certificate, this attribute allows the
-	// caller to provide multiple certificates that support different
-	// algorithms. The final certificate will be selected based on the DTLS
-	// handshake, which establishes which certificates are allowed. The
-	// PeerConnection implementation selects which of the certificates is
-	// used for a given connection; how certificates are selected is outside
-	// the scope of this specification. If this value is absent, then a default
-	// set of certificates is generated for each PeerConnection instance.
-	Certificates []Certificate
+	// Certificates are not supported in the JavaScript/Wasm bindings.
+	// Certificates []Certificate
 
 	// ICECandidatePoolSize describes the size of the prefetched ICE pool.
 	ICECandidatePoolSize uint8
