+++ examples/data-channels/jsfiddle/main.go
@@ -1,4 +1,4 @@
-// +build js
+// +build js,wasm
 
 package main
